requirements.txt
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
contourpy==1.2.0
cycler==0.12.1
darr==0.5.5
EasyProcess==1.1
entrypoint2==1.1
fonttools==4.48.1
h11==0.14.0
idna==3.4
joblib==1.3.2
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.8.2
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.0.0
PyAutoGUI==0.9.54
pycparser==2.21
pygame==2.3.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
requests==2.28.2
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sound==0.1.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
threadpoolctl==3.2.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.15
wsproto==1.2.0
yellowbrick==1.5